+++ src/client/cypress/e2e/app.cy.js
@@ -1,4 +1,4 @@
-import { loginAsAdmin, loginAsEditorInViewerMode } from "./helpers/testHelpers";
+import { loginAsEditor } from "./helpers/testHelpers";
 
 describe("General app tests", () => {
   it("Displays the login page in the correct language", () => {
@@ -25,26 +25,8 @@ describe("General app tests", () => {
   });
 
   it("Displays the current host as app title", () => {
-    loginAsEditorInViewerMode();
+    loginAsEditor();
     cy.visit("/");
     cy.get('[data-cy="app-title"]').contains("localhost");
-  });
-
-  it("Correctly navigates back and forth from settings", () => {
-    // correctly navigate back to viewer mode
-    loginAsAdmin();
-    cy.visit("/");
-    cy.get('[data-cy="menu"]').click();
-    cy.get('[data-cy="settings-list-item"]').click();
-    cy.contains("h3", "Done").click();
-    cy.contains("span", "Viewer");
-
-    // correctly navigate back to editor mode
-    cy.get('[data-cy="menu"]').click();
-    cy.get('[data-cy="editor-list-item"]').click();
-    cy.get('[data-cy="menu"]').click();
-    cy.get('[data-cy="settings-list-item"]').click();
-    cy.contains("h3", "Done").click();
-    cy.contains("span", "Editor");
   });
 });