+++ src/client/src/commons/form/dateField.jsx
@@ -48,6 +48,7 @@ class DateField extends React.Component {
               iconPosition="left"
               placeholder={placeholder}
               spellCheck="false"
+              data-cy="datepicker"
               className="datepicker-input"
               value={moment(this.state.date).isValid() ? moment(this.state.date).format("DD.MM.YYYY") : ""}
               disabled={true}