+++ src/client/src/commons/menu/menuContainer.jsx
@@ -5,27 +5,16 @@ import MenuComponent from "./menuComponent";
 
 class MenuContainer extends React.Component {
   render() {
-    const { history, location } = this.props;
+    const { history } = this.props;
     return (
       <MenuComponent
         handleModeChange={mode => {
           if (mode === "editor") {
             history.push(`/editor`);
-          } else if (mode === "viewer") {
-            history.push("/");
           } else if (mode.indexOf("setting") >= 0) {
             history.push(`/${mode}`);
           }
-        }}
-        mode={(() => {
-          if (location.pathname.indexOf("setting/") >= 0) {
-            return "setting";
-          } else if (location.pathname.indexOf("editor") >= 0) {
-            return "editor";
-          } else {
-            return "viewer";
-          }
-        })()}>
+        }}>
         {this.props.children}
       </MenuComponent>
     );
@@ -36,9 +25,6 @@ MenuContainer.propTypes = {
   children: PropTypes.node,
   history: PropTypes.shape({
     push: PropTypes.func,
-  }).isRequired,
-  location: PropTypes.shape({
-    pathname: PropTypes.string,
   }).isRequired,
 };