+++ src/client/src/stratigraphy/index.js
@@ -1,1 +1,0 @@
-export { default as Stratigraphy } from "./stratigraphy";